feat: no connect window

+++ flutter/lib/desktop/pages/server_page.dart
@@ -55,14 +55,18 @@ class _DesktopServerPageState extends State<DesktopServerPage>
 
   @override
   void onWindowClose() {
-    Future.wait([gFFI.serverModel.closeAll(), gFFI.close()]).then((_) {
-      if (isMacOS) {
-        RdPlatformChannel.instance.terminate();
-      } else {
-        windowManager.setPreventClose(false);
-        windowManager.close();
-      }
-    });
+    if (gFFI.serverModel.clients.isNotEmpty) {
+      hideCurrentCmWindow();
+    } else {
+      gFFI.close().then((_) {
+        if (isMacOS) {
+          RdPlatformChannel.instance.terminate();
+        } else {
+          windowManager.setPreventClose(false);
+          windowManager.close();
+        }
+      });
+    }
     super.onWindowClose();
   }
 
@@ -164,14 +168,6 @@ class ConnectionManagerState extends State<ConnectionManager>
   @override
   Widget build(BuildContext context) {
     final serverModel = Provider.of<ServerModel>(context);
-    pointerHandler(PointerEvent e) {
-      if (serverModel.cmHiddenTimer != null) {
-        serverModel.cmHiddenTimer!.cancel();
-        serverModel.cmHiddenTimer = null;
-        debugPrint("CM hidden timer has been canceled");
-      }
-    }
-
     return serverModel.clients.isEmpty
         ? Column(
             children: [
@@ -183,10 +179,7 @@ class ConnectionManagerState extends State<ConnectionManager>
               ),
             ],
           )
-        : Listener(
-            onPointerDown: pointerHandler,
-            onPointerMove: pointerHandler,
-            child: DesktopTab(
+        : DesktopTab(
               showTitle: false,
               showMaximize: false,
               showMinimize: true,
@@ -261,8 +254,7 @@ class ConnectionManagerState extends State<ConnectionManager>
                   );
                 },
               ),
-            ),
-          );
+            );
   }
 
   Widget buildSidePage() {
@@ -324,25 +316,16 @@ class ConnectionManagerState extends State<ConnectionManager>
   }
 
   Future<bool> handleWindowCloseButton() async {
-    var tabController = gFFI.serverModel.tabController;
-    final connLength = tabController.length;
-    if (connLength <= 1) {
-      windowManager.close();
-      return true;
-    } else {
-      final bool res;
-      if (!option2bool(kOptionEnableConfirmClosingTabs,
-          bind.mainGetLocalOption(key: kOptionEnableConfirmClosingTabs))) {
-        res = true;
-      } else {
-        res = await closeConfirmDialog();
-      }
-      if (res) {
-        windowManager.close();
-      }
-      return res;
-    }
+    await hideCurrentCmWindow();
+    return false;
   }
+}
+
+Future<void> hideCurrentCmWindow() async {
+  await windowManager.setOpacity(0);
+  bind.mainHideDock();
+  await windowManager.minimize();
+  await windowManager.hide();
 }
 
 Widget buildConnectionCard(Client client) {
@@ -389,7 +372,7 @@ class _CloseButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return IconButton(
       onPressed: () {
-        windowManager.close();
+        hideCurrentCmWindow();
       },
       icon: const Icon(
         IconFont.close,
